Modified wiggle.cpp & wiggle.h for more accurate floating point number arithmetics. Modified rsem-gen-transcript-plots to output warning messages instead of exit the script when positions where multiread track has less coverage than uniq track are detected. Thanks Han Lin for pointing out the problem and suggesting possible fixes

+++ wiggle.h
@@ -10,7 +10,7 @@ extern bool no_fractional_weight; // if no_frac_weight == true, each alignment c
 
 struct Wiggle {
     std::string name;
-    std::vector<float> read_depth;
+    std::vector<double> read_depth;
     size_t length;
 };
